fix(upload): check if path exists before write

+++ www/cgi-bin/upload.py
@@ -14,19 +14,15 @@ split_input = input_data.split(boundary_b)
 
 for content in split_input:
 	if b"filename" in content:
-		# sys.stderr.write("ok1\n")
 		filename = content.split(b"filename=\"")[1].split(b"\"")[0].decode()
 		file_content = content.split(b"\r\n\r\n")[1]
-
-		# file_path = '.' + os.path.join(os.path.dirname(__file__), filename)
-		# os.makedirs(os.path.dirname(file_path), exist_ok=True)
-		with open(upload_location + filename, 'wb') as file:
-			file.write(file_content)
-
-		body = "File succesfully uploaded !"
+		if os.path.exists(upload_location):
+			with open(upload_location + filename, 'wb') as file:
+				file.write(file_content)
+				body = "File succesfully uploaded!"
+		else:
+			body = "Invalid upload save location!"
 		body_length = len(body)
-
-		# sys.stderr.write(f"{body}\r")
 		print("Content-Type: text/plain\r")
 		print(f"Content-Length: {body_length}\r\n\r")
 		print(f"{body}")
